move plotters back to main package
