+++ Cargo.toml
@@ -34,6 +34,7 @@ ark-serialize = { version = "0.5.0", features = ["derive"] }
 bincode = "1.3"
 
 inkwell = { version = "0.5", features = ["llvm18-0"] }
+plotters = { version = "0.3.5" }
 wasmtime = { version = "31" }
 
 tracing = "0.1.41"
