+++ src/lib.rs
@@ -2,6 +2,7 @@ pub mod abi_helpers;
 pub mod compiler;
 pub mod driver;
 pub mod error;
+pub mod plotting;
 pub mod project;
 
 pub use mavros_artifacts as artifacts;
