+++ vm/Cargo.toml
@@ -9,6 +9,5 @@ ark-ff = { version = "0.5.0" }
 mavros-artifacts = { path = "../mavros-artifacts" }
 noirc_abi = { git = "https://github.com/noir-lang/noir.git", branch = "master", package = "noirc_abi" }
 opcode-gen = { path = "../opcode-gen" }
-plotters = { version = "0.3.5" }
 serde = { version = "1.0.204", features = ["derive"] }
 tracing = "0.1.41"
